+++ app/app.js
@@ -1,7 +1,9 @@
-var app = angular.module('test', [
-        "firebase",
-        "ui.router"
-    ])
+'use strict';
+
+module.exports = angular.module('fire-emblem', [
+    "firebase",
+    "ui.router"
+])
     .config(function($locationProvider, $urlRouterProvider) {
         $locationProvider.html5Mode({
             enabled: true
@@ -11,4 +13,7 @@ var app = angular.module('test', [
     })
     .run(function(characters, classes, skills) {
         console.debug('initializing services...');
+    })
+    .factory('_', function() {
+        return require('underscore');
     });
