+++ app/components/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+require('./character');
