+++ app/services/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+angular.module('fire-emblem')
+    .service('firebase', require('./firebase'))
+    .service('characters', require('./characters'))
+    .service('classes', require('./classes'))
+    .service('skills', require('./skills'));
