--- app/states/characters/characterState.js
@@ -1,8 +1,0 @@
-app.config(function($stateProvider) {
-    $stateProvider
-        .state('root.characters', {
-            abstract: true,
-            template: '<ui-view/>',
-            url: "/characters"
-        });
-});
