--- app/states/characters/details/charactersDetailsState.js
@@ -1,7 +1,0 @@
-app.config(function($stateProvider) {
-    $stateProvider
-        .state('root.characters.details', {
-            url: "/name/:name",
-            templateUrl: "/app/states/characters/characters.html"
-        });
-});
