--- app/states/characters/list/charactersListState.js
@@ -1,10 +1,0 @@
-app.config(function($stateProvider) {
-    $stateProvider
-        .state('root.characters.list', {
-            url: "/list",
-            templateUrl: "/app/states/characters/list/charactersList.html",
-            controller: function($scope, characters) {
-                $scope.characters = characters.all;
-            }
-        });
-});
