--- app/states/home/homeState.js
@@ -1,7 +1,0 @@
-app.config(function($stateProvider) {
-    $stateProvider
-        .state('root.home', {
-            url: "/",
-            templateUrl: "app/states/home/home.html"
-        });
-});
