+++ app/states/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+require('./rootState');
+require('./home');
+require('./characters');
