--- bower.json
@@ -1,23 +1,0 @@
-{
-  "name": "firebaseEmblem",
-  "version": "0.0.0",
-  "homepage": "https://github.com/vitorfranca/firebaseEmblem",
-  "authors": [
-    "Vitor França <vitor.franca@izzui.com>"
-  ],
-  "license": "MIT",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "dependencies": {
-    "angular": "~1.4.1",
-    "firebase": "~2.2.7",
-    "angularfire": "~1.1.2",
-    "angular-ui-router": "~0.2.15",
-    "underscore": "~1.8.3"
-  }
-}
