+++ index.html
@@ -7,12 +7,7 @@
 
 <base href="/" />
 
-<script src="bower_components/angular/angular.min.js"></script>
-<script src="bower_components/underscore/underscore-min.js"></script>
-<script src="bower_components/angular-ui-router/release/angular-ui-router.min.js"></script>
-
-<script src="bower_components/firebase/firebase.js"></script>
-<script src="bower_components/angularfire/dist/angularfire.min.js"></script>
+<script src="assets/js/vendor.js"></script>
 
 <link rel="stylesheet" href="assets/css/base.css" media="screen" title="no title" charset="utf-8">
 
@@ -20,11 +15,13 @@
 </head>
 
 <body>
-  <div ng-app='test'>
+  <div ng-app='fire-emblem'>
     <div ui-view></div>
   </div>
 
-  <script src="assets/js/app.js"></script>
+  <!-- <script src="assets/js/app.js"></script> -->
+  <script src="assets/js/bundle.js"></script>
+
 
 </body>
 </html>
