+++ main.js
@@ -1,0 +1,6 @@
+'use strict';
+
+require('./app/app.js');
+require('./app/services');
+require('./app/states');
+require('./app/components');
