+++ vendor.js
@@ -1,0 +1,7 @@
+'use strict';
+
+var angular = require('angular');
+var uiRouter = require('ui-router');
+
+var firebase = require('firebase');
+var angularfire = require('angularfire');
